backend/django/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.14
gunicorn==20.0.4
python-dotenv==0.13.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
importlib-metadata==4.12.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
pathspec==0.10.3
platformdirs==2.6.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.10
wrapt==1.14.1
